Create hangman game
The Hangman Game is a C language project that is quite simple in nature. It shows that you can also develop games using C. The game feels very interesting to the user because of the coding.You will have to guess one character when you run the game and you will be given five chances to do this.